--- _sebelib/__init__.py
@@ -1,8 +1,0 @@
-
-## 1/6/2020 16:50 - "Don't pospone something just because you have enough time"
-
-        
-## page name and url for render template
-class PageCtx:
-    def __init__(self, name, url):
-        self.name = "%-10s"%name; self.url = url
