+++ _static/html/base.html
@@ -31,15 +31,15 @@
 
             <!-- brand name  -->
             <a class='navbar-brand' rel="nofollow" href='/'>
-              {% if brand_name %}{{ brand_name }} {% else %} SEBE RestApi {% endif %}
+              {% if brand_name %}{{ brand_name }} {% else %} {{API_TITLE}} RestApi {% endif %}
             </a>
 
              <!-- login logout  -->
             <ul class="nav navbar-nav pull-right">  
               {% if user.is_authenticated %}
-              <li><a href="{% url 'accounts-logout' %}">Log out</a></li>
+              <li><a href="{% url 'api-logout' %}">Log out</a></li>
               {% else %}
-              <li><a href="{% url 'accounts-login' %}">Log in</a></li>
+              <li><a href="{% url 'api-login' %}">Log in</a></li>
               {% endif %}
             </ul>
