+++ _static/html/login.html
@@ -1,0 +1,43 @@
+{% extends "rest_framework/base.html" %}
+{% load rest_framework %}
+
+{% block body %}
+<body class="container">
+  <div class="container-fluid" style="margin-top: 30px">
+    <div class="row-fluid">
+      <div class="well" style="width: 320px; margin-left: auto; margin-right: auto">
+
+        
+        <div class="row-fluid">
+          <div>
+            <h3 style="margin: 0 0 20px; text-align:center">{{ title }} Authentication</h3>
+          </div>
+        </div>
+
+        {% if error_message %}
+      <p style="text-align:center; color: rgb(179, 21, 21);"> {{error_message}} </p>
+        {% endif %}
+
+        <div class="row-fluid">
+          <div>
+            <form action="" method="POST">
+                {% csrf_token %}
+                
+                <label>Username:</label>
+                <input class="form-control textinput textInput" type="text" name="username" value="{{ username }}" required autofocus>
+                <br>
+                <label>Password:</label>
+                <input type="password" name="password" class="form-control textinput textInput" required>
+                <br>
+                <div class="form-actions-no-box">
+                <input type="submit" name="submit" value="Log in" class="btn btn-primary form-control" id="submit-id-submit">
+                </div>
+            </form>
+
+          </div>
+        </div><!-- /.row-fluid -->
+      </div><!--/.well-->
+    </div><!-- /.row-fluid -->
+  </div><!-- /.container-fluid -->
+</body>
+{% endblock %}
