+++ _static/html/objects.html
@@ -3,7 +3,7 @@
 
 {% block content %}
 
-<h1 style="text-align: center; margin-top:10px;margin-bottom:20px">{{title}}</h1>
+<h1 style="text-align: center; margin-top:10px;margin-bottom:20px">{{model_name}}</h1>
 
 <pre class="prettyprint">
     {
