+++ _static/html/rest_framework/api.html
@@ -1,15 +1,15 @@
 {% extends "rest_framework/base.html" %}
 
 {% block branding %}
-<span><a href="/" class="navbar-brand">SEBE RestApi</a></span>
+<span><a href="/" class="navbar-brand">{{API_TITLE}} RestApi</a></span>
 {% endblock %}
 
 {% block userlinks %}
 
 {% if user.is_authenticated %}
-<li><a href="{% url 'accounts-logout' %}">Log out</a></li>
+<li><a href="{% url 'api-logout' %}">Log out</a></li>
 {% else %}
-<li><a href="{% url 'accounts-login' %">Log in</a></li>
+<li><a href="{% url 'api-login' %}">Log in</a></li>
 {% endif %}
 
 {% endblock %}
