+++ documents/urls.py
@@ -1,19 +1,11 @@
 from django.urls import path
-from _sebelib.templates import register_models, AppHomeResponse
+from drfvg import register_models
 
 ## models to register
 from .models.technical_sheet import TechnicalSheet
 
-## register models
-API_MODELS_REGITRY = [
-    TechnicalSheet
-]
-APP_NAME = 'documents'
-
 ## documents/
-urlpatterns = [
-    path('', AppHomeResponse(API_MODELS_REGITRY), name='%s-home'%APP_NAME),
-] + register_models(API_MODELS_REGITRY)
+urlpatterns = [ ] + register_models( [ TechnicalSheet ], app_name='documents')
